normalizer: prevent memoization on normalization requests
A set_memo call will set the value of a reference in the machine's
environment, which can leak from the normalization request into the
enclosing `norm` invocation. This is wrong since we may be using a
completely different configuration between the two situations.

It may be more efficient to copy the references so the underlying norm
request can also make use of memoization.